A few small things...
carried over code for implementation of vgg16.

fixed an issue with a missing file causing code not to run.

halfed the batch size for training, testing, and validation as to not exceed memory constraints.

commented out code that disabled running on the GPU.

changed the patience for model training from 30 to 5 for now, to debug code quicker.